Validating user/shared seeds correspond to xpubkeys
We did not do any checking before whether the password the user provided
resulted in a valid decrypted seed that corresponds to one of the
xpubkeys.

We also did not check if the unencrypted user seed corresponds to one of
the xpubkeys.

This resulted in a bad UX whereas user error would result in weird
errors that seemed like a problem in the code.
