+++ lib/multisig/transaction.js
@@ -3,8 +3,10 @@
 // Dependencies.
 var Bitcoin = require('bitcoinjs-lib');
 var Buffer = require('buffer');
+
+var generateHdWallet = require('./generateHdWallet');
 var constants = require('./constants');
-var utils = require('./utils')
+var utils = require('./utils');
 
 /***
 Transaction builder */
@@ -51,27 +53,13 @@ Transaction.prototype = {
   addOutput: function() {
     this.tb.addOutput(this.destinationAddress, this.destinationAmount);
   },
-  generateHdWallet: function(seed_or_xprvkey) {
-    // Dealing with a seed.
-    if (seed_or_xprvkey.match(/^(L|K)[a-zA-Z0-9]*$/)) {
-      var key = Bitcoin.ECKey.fromWIF(seed_or_xprvkey);
-      return Bitcoin.HDNode.fromSeedBuffer(key.d.toBuffer());
-    }
-
-    // Dealing with a extended private key.
-    if (seed_or_xprvkey.match(/^xprv[a-zA-Z0-9]*$/)) {
-      return Bitcoin.HDNode.fromBase58(seed_or_xprvkey);
-    }
-
-    return false;
-  },
   signInputs: function() {
     var that = this;
 
     // Generate the HD wallets.
     this.hdWallets = [
-      this.generateHdWallet(this.seeds.shared),
-      this.generateHdWallet(this.seeds.user)
+      generateHdWallet(this.seeds.shared),
+      generateHdWallet(this.seeds.user)
     ];
 
     // Generate a random thing to sign.
