+++ lib/multisig/vault.js
@@ -1,9 +1,11 @@
 "use strict";
 
 var Bitcoin = require('bitcoinjs-lib');
+
 var Address = require('./address');
 var Transaction = require('./transaction');
 var Bip38Worker = require('./bip38worker');
+var generateHdWallet = require('./generateHdWallet');
 var constants = require('./constants');
 var utils = require('./utils');
 
@@ -65,6 +67,23 @@ Vault.prototype = {
       address.getUnspentOutputs().done(this.updateBalance.bind(this));
     }
   },
+  xpubkeysIncludeSeed: function(seed) {
+    var node = generateHdWallet(seed);
+    if (!node) {
+      return false;
+    }
+
+    var xpub = node.neutered().toBase58();
+    if (!xpub) {
+      return false;
+    }
+
+    if (!this.xpubkeys.length) {
+      return false;
+    }
+
+    return this.xpubkeys.filter(function(xpk) { return xpk[0] == xpub; }).length > 0;
+  },
   buildTransaction: function(options) {
     this.tx = new Transaction({
       addresses: this.addresses,
@@ -76,6 +95,12 @@ Vault.prototype = {
     });
 
     this.tx.onReady = options.onReady;
+
+    // Validate user key.
+    if (!this.xpubkeysIncludeSeed(this.tx.seeds.user)) {
+      var err = new Error('The user seed does not correspond to a xpubkey.');
+      this.tx.onReady(err);
+    }
 
     // Build transaction.
     this.tx.addInputs();
@@ -92,6 +117,12 @@ Vault.prototype = {
     // Use the just decrypted seed.
     this.tx.seeds.shared = shared_seed;
 
+    // Validate decrypted shared key.
+    if (!this.xpubkeysIncludeSeed(shared_seed)) {
+      var err = new Error('The shared encrypted seed password is incorrect.');
+      this.tx.onReady(err);
+    }
+
     // Sign transaction.
     this.tx.signInputs();
 
@@ -101,6 +132,6 @@ Vault.prototype = {
     // Call ready callback.
     this.tx.onReady();
   }
-}
+};
 
 module.exports = Vault;
