+++ lib/tool.js
@@ -135,7 +135,12 @@ $(function() {
         percent = cur_percent;
         $('#decrypt_percent').text(percent);
       },
-      onReady: function() {
+      onReady: function(err) {
+        if (err) {
+          alert(err.message);
+          throw err;
+        }
+
         $('.fees_btc').text(vault.tx.getMinerFee());
         $('.total_btc').text(vault.tx.getDestinationAmount());
         $('.raw code').text(vault.tx.transaction.toHex());
